# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Domi API
  version: 1.1.0
  description: Versioned contract shared by Domi clients and the Rails API.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/v1/auth/email-verification:
    post:
      operationId: verifyEmail
      tags: [Authentication]
      requestBody:
        $ref: "#/components/requestBodies/Token"
      responses:
        "200": { $ref: "#/components/responses/Authenticated" }
        "422": { $ref: "#/components/responses/Error" }
...
  /api/v1/households/{householdId}/ownership:
    parameters:
      - $ref: "#/components/parameters/HouseholdId"
    post:
      operationId: transferOwnership
      tags: [Households]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [membershipId]
              properties:
                membershipId: { type: string, format: uuid }
      responses:
        "200":
          description: Ownership transferred.
          content:
            application/json:
              schema:
                type: object
                required: [membership]
                properties:
                  membership: { $ref: "#/components/schemas/Membership" }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/households/{householdId}/invitations:
    parameters:
      - $ref: "#/components/parameters/HouseholdId"
    get:
      operationId: listInvitations
      tags: [Households]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Household invitations without secret tokens.
          content:
            application/json:
              schema:
                type: object
                required: [invitations]
                properties:
                  invitations:
                    type: array
                    items: { $ref: "#/components/schemas/Invitation" }
    post:
      operationId: createInvitation
      tags: [Households]
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Invitation created; token is returned only once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationSecret"
        "403": { $ref: "#/components/responses/Error" }
...
  /api/v1/invitations/accept:
    post:
      operationId: acceptInvitation
      tags: [Households]
      security: [{ bearerAuth: [] }]
      requestBody:
        $ref: "#/components/requestBodies/Token"
      responses:
        "201": { $ref: "#/components/responses/Household" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    HouseholdId:
      name: householdId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  requestBodies:
...
    Token:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [token]
            properties:
              token: { type: string }
...
  responses:
    Authenticated:
      description: Authenticated user and new mobile credentials.
      content:
        application/json:
          schema:
            type: object
            required: [user, session]
            properties:
              user: { $ref: "#/components/schemas/User" }
              session: { $ref: "#/components/schemas/Session" }
...
    Household:
      description: Household visible to the current user.
      content:
        application/json:
          schema:
            type: object
            required: [household]
            properties:
              household: { $ref: "#/components/schemas/Household" }
    Error:
      description: Request failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    User:
      type: object
      additionalProperties: false
      required: [id, email, displayName, emailVerified, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        emailVerified: { type: boolean }
        createdAt: { type: string, format: date-time }
    Session:
      type: object
      additionalProperties: false
      required: [accessToken, refreshToken, accessExpiresAt, refreshExpiresAt]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        accessExpiresAt: { type: string, format: date-time }
        refreshExpiresAt: { type: string, format: date-time }
    Household:
      type: object
      additionalProperties: false
      required: [id, name, timezone, role, version, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        timezone: { type: string }
        role: { type: string, enum: [owner, member] }
        version: { type: integer }
        createdAt: { type: string, format: date-time }
    Membership:
      type: object
      additionalProperties: false
      required: [id, role, joinedAt, user]
      properties:
        id: { type: string, format: uuid }
        role: { type: string, enum: [owner, member] }
        joinedAt: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/User" }
    Invitation:
      type: object
      required: [id, expiresAt, createdAt, createdBy]
      properties:
        id: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time }
        revokedAt: { type: [string, "null"], format: date-time }
        acceptedAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        createdBy: { $ref: "#/components/schemas/User" }
    InvitationSecret:
      type: object
      required: [invitation, token, inviteUrl]
      properties:
        invitation: { $ref: "#/components/schemas/Invitation" }
        token: { type: string }
        inviteUrl: { type: string }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId, details]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: object, additionalProperties: true }